move choicetree.py to shrinker/
