+++ hypothesis-python/src/hypothesis/internal/conjecture/shrinking/choicetree.py
